requirements.txt
Flask==3.0.3
Flask-Assets==2.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-WTF==1.2.1
SQLAlchemy-Searchable==2.1.0
environs==11.0.0
cssutils==2.10.2
jsmin==3.0.1
psycopg2==2.9.9
requests==2.31.0